requirements.txt
Django
celery
redis
python-dotenv
ckeditor
Pillow
django-filter
django-allauth
requests
